feat(toolkit-cleaner): allow only manual runs (#160)
Thank you for writing this toolkit clean-up logic. It's a great solution & I'd love to see it integration into the core CDK project at some point. This will really help out several of my larger clients that rely heavily on CDK.

In the meantime, I tested this out against my personal stacks and it worked flawlessly. However, I'd like to be able to only run it manually (e.g., not on a schedule) for this account. This PR adds the ability to pass `schedule: null`, which skips creating the eventbridge rule.